[UPUP-1033]: fix comments
