+++ Sources/MUDKit/UI/FileSystem/Views/WebView.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+import WebKit
+
+struct WebView: UIViewRepresentable {
+    let url: URL
+    
+    func makeUIView(context: Context) -> WKWebView {
+        let webView = WKWebView()
+        webView.loadFileURL(url, allowingReadAccessTo: url.deletingLastPathComponent())
+        
+        return webView
+    }
+    
+    func updateUIView(_ uiView: WKWebView, context: Context) {}
+}
